# swagger.yaml
openapi: 3.0.4
info:
  title: Paved Way API
  description: API per la gestione degli eventi sulla mobilità della città di Trento.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Server locale per lo sviluppo e il test.
...
paths:
...
  /users/{id}/reports:
    get:
      tags:
        - Utenti
      summary: Restituisce la lista delle segnalazioni di un utente
      description: Restituisce la lista delle segnalazioni di un utente specifico in base all'ID fornito.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID dell'utente di cui restituire le segnalazioni.
        - in: query
          name: latitude
          schema:
            type: float
          description: Latitudine del punto attorno a cui si vuole ricercare dei cantieri.
        - in: query
          name: longitude
          schema:
            type: float
          description: Longitudine del punto attorno a cui si vuole ricercare dei cantieri.
        - in: query
          name: radius
          schema:
            type: integer
          description: Raggio del cerchio con centro il punto attorno a cui si vuole ricercare dei cantieri.
        - in: query
          name: offset
          schema:
            type: integer
          description: Numero di segnalazioni da saltare prima di restituire i risultati.
        - in: query
          name: limit
          schema:
            type: integer
          description: Numero massimo di segnalazioni da restituire.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        "404":
          description: Utente non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /reports:
    get:
      tags:
        - Segnalazioni
      summary: Restituisce la lista delle segnalazioni
      description: Restituisce la lista delle segnalazioni.
      parameters:
        - in: query
          name: now
          schema:
            type: boolean
          description: >-
            Se fornito, filtra le segnalazioni in corso nell'ora attuale.
            Il filtro è applicato se il parametro è "true".
            Il filtro è ignorato se il parametro è "false" oppure non è un valore booleano.
            Se il parametro è "true", non può essere usato insieme a "date".
        - in: query
          name: date
          schema:
            oneOf:
              - type: string
                format: date
              - type: string
                format: date-time
          description: >-
            Se fornito, filtra le segnalazioni in corso nella data fornita.
            Non può essere usato insieme a "now".
        - in: query
          name: latitude
          schema:
            type: number
            format: double
          description: Latitudine del punto attorno a cui si vuole ricercare dei cantieri.
        - in: query
          name: longitude
          schema:
            type: number
            format: double
          description: Longitudine del punto attorno a cui si vuole ricercare dei cantieri.
        - in: query
          name: radius
          schema:
            type: integer
          description: Raggio del cerchio con centro il punto attorno a cui si vuole ricercare delle segnalazioni.
        - in: query
          name: offset
          schema:
            type: integer
          description: Numero di segnalazioni da saltare prima di restituirle.
        - in: query
          name: limit
          schema:
            type: integer
          description: Numero massimo di segnalazioni da restituire.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        "400":
          description: Richiesta non valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - Segnalazioni
      summary: Crea una nuova segnalazione
      description: Crea una nuova segnalazione. Devi essere loggato per poterne creare una nuova.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Report"
      responses:
        "201":
          description: Report creato con successo.
        "401":
          description: Non autorizzato. Devi fornire una chiave API valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /reports/{id}:
    get:
      tags:
        - Segnalazioni
      summary: Restituisce una segnalazione specifica
      description: Restituisce una segnalazione specifica in base all'ID fornito.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID della segnalazione da restituire.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Segnalazione non trovata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - Segnalazioni
      summary: Elimina una segnalazione specifica.
      description: >-
        Elimina una segnalazione specifica in base all'ID fornito. 
        Devi essere autenticato per eliminare una segnalazione.
        Se sei autenticato come amministratore, puoi eliminare qualsiasi segnalazione.
        Se sei autenticato come cittadino, puoi eliminare solo le segnalazioni che hai creato.
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID della segnalazione da eliminare.
      responses:
        "204":
          description: Segnalazione eliminata con successo.
        "401":
          description: Non autorizzato. Devi fornire una chiave API valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: >-
            Accesso negato. 
            Devi essere un amministratore per eliminare una segnalazione qualsiasi o
            un cittadino per eliminare solo le segnalazioni che hai creato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Segnalazione non trovata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags:
        - Segnalazioni
      summary: Modifica una segnalazione specifica.
      description:
        Permette la modifica di una segnalazione. Per modificare una segnalazione
        è necessario essere autenticati come cittadino e averla creata.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneof:
                - $ref: "#/components/schemas/UpdateReportCitizen"
                - $ref: "#/components/schemas/UpdateReportAdmin"
      responses:
        "204":
          description: Segnalazione modificata con successo.
        "401":
          description: Non autorizzato. Devi fornire una chiave API valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: >-
            Accesso negato.
            Devi essere un amministratore per modificare questo campo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Segnalazione non trovata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /reports/{id}/comments:
    get:
      tags:
        - Segnalazioni
      summary: Restituisce la lista dei commenti di una segnalazione
      description: >-
        Restituisce la lista dei commenti di una segnalazione specifica in base all'ID fornito.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID della segnalazione di cui restituire i commenti.
        - in: query
          name: offset
          schema:
            type: integer
          description: Numero di commenti da saltare prima di restituire i risultati.
        - in: query
          name: limit
          schema:
            type: integer
          description: Numero massimo di commenti da restituire.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "404":
          description: Segnalazione non trovata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Segnalazioni
      summary: Aggiunge un commento a una segnalazione
      description: >-
        Permette di aggiungere un commento a una segnalazione specifica.
        Devi essere autenticato per aggiungere un commento.
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID della segnalazione a cui aggiungere il commento.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: Testo del commento.
      responses:
        "201":
          description: Commento aggiunto con successo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Richiesta non valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Non autorizzato. Devi fornire una chiave API valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Segnalazione non trovata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Chiave API per l'autenticazione.
        Deve essere fornita nell'intestazione della richiesta.
        Deve essere un token JWT valido.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Messaggio di errore.
        code:
          type: integer
          description: Codice di errore.
        message:
          type: string
          description: Messaggio di errore dettagliato.
...
    Location:
      type: object
      required:
        - latitude
        - longitude
        - street
        - number
        - city
        - code
      properties:
        latitude:
          type: number
          format: float
          description: Latitudine della posizione.
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: float
          description: Longitudine della posizione.
          minimum: -180
          maximum: 180
        street:
          type: string
          description: Nome della strada.
        number:
          type: string
          description: Numero civico.
        city:
          type: string
          description: Nome della città.
        code:
          type: string
          minLength: 5
          maxLength: 5
          description: Codice di Avviamento Postale.
    Duration:
      type: object
      required:
        - start
      properties:
        start:
          type: string
          format: date-time
          description: Data e ora di inizio dell'evento.
        end:
          type: string
          format: date-time
          description: Data e ora di fine dell'evento.
    Comment:
      type: object
      required:
        - text
      properties:
        id:
          type: integer
          description: ID del commento (generato automaticamente dal server).
        userId:
          type: integer
          description: ID dell'utente che ha scritto il commento.
        text:
          type: string
          description: Testo del commento.
        createdAt:
          type: string
          format: date-time
          description: Data e ora di creazione del commento.
    Event:
      type: object
      required:
        - name
        - info
        - location
        - duration
      properties:
        id:
          type: integer
          description: ID dell'evento (generato automaticamente dal server).
        name:
          type: string
          description: Nome dell'evento.
        info:
          type: string
          description: Descrizione dell'evento.
        location:
          allOf:
            - $ref: "#/components/schemas/Location"
          description: Posizione dell'evento.
        duration:
          allOf:
            - $ref: "#/components/schemas/Duration"
          description: Durata dell'evento.
        createdAt:
          type: string
          format: date-time
          description: Data e ora di creazione dell'evento (ora attuale se non specificata).
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
          description: Lista dei commenti associati all'evento.
    Report:
      allOf:
        - $ref: "#/components/schemas/Event"
        - type: object
          properties:
            createdBy:
              type: string
              description: ID dell'utente che ha creato la segnalazione.
            photos:
              type: array
              items:
                type: string
                format: uri
              description: Lista degli URL delle foto associate alla segnalazione.
            rating:
              type: integer
              description: Punteggio della segnalazione (impostato a 0 per una nuova segnalazione).
            status:
              type: string
              enum:
                - pending
                - approved
                - rejected
                - solved
              description: Stato della segnalazione.
    UpdateReportCitizen:
      type: object
      properties:
        name:
          type: string
          description: Nome dell'evento.
        info:
          type: string
          description: Descrizione dell'evento.
        duration:
          allOf:
            - $ref: "#/components/schemas/Duration"
          description: Durata dell'evento.
        photos:
          type: array
          items:
            type: string
            format: uri
          description: Lista degli URL delle foto associate alla segnalazione.
        status:
          type: string
          enum:
            - solved
          description: Stato della segnalazione.
    UpdateReportAdmin:
      type: object
      properties:
        status:
          type: string
          enum:
            - pending
            - approved
            - rejected
            - solved
          description: Stato della segnalazione.
...